+++ assets/scripts.js
@@ -4,20 +4,23 @@ function init(keys) {
     let append = '';
     for (let i = 0; i < keys.length; i++) {
         append += '<tr id="monitor-' + keys[i].id + '">' +
-            '<td class="align-middle text-nowrap"><span class="status"></span> ' + keys[i].name + '</td>' +
+            '<th class="align-middle text-nowrap"><span class="status"></span> ' + keys[i].name + '</th>' +
             '<td class="align-middle text-center latest">' +
-            '<span class="time"></span>' +
-            '<span class="reason"></span>' +
-            '<span class="duration"></span></td>' +
-            '<td class="align-middle text-center"><span class="r0"></span></td>' +
-            '<td class="align-middle text-center"><span class="r1"></span></td>' +
-            '<td class="align-middle text-center"><span class="r2"></span></td>' +
-            '<td class="align-middle text-center"><span class="r3"></span></td>' +
+            '<span class="time">loading...</span>' +
+            '<span class="reason">loading...</span>' +
+            '<span class="duration">loading...</span></td>' +
+            '<td class="align-middle text-center"><span class="r0">loading...</span></td>' +
+            '<td class="align-middle text-center"><span class="r1">loading...</span></td>' +
+            '<td class="align-middle text-center"><span class="r2">loading...</span></td>' +
+            '<td class="align-middle text-center"><span class="r3">loading...</span></td>' +
             '</tr>'
         ;
     }
 
     $('table').append(append);
+
+    loadData();
+    checkStatus();
 }
 
 let apiKeys = null;
@@ -33,7 +36,14 @@ $.ajax({
 
 function formatDatetime(timestamp) {
     let date = new Date(timestamp * 1000);
-    let options = {year: 'numeric', month: 'short', day: '2-digit', hour: '2-digit', minute: '2-digit'};
+    let options = {
+        year: 'numeric',
+        month: 'short',
+        day: '2-digit',
+        hour: '2-digit',
+        minute: '2-digit',
+        second: '2-digit'
+    };
     return new Intl.DateTimeFormat('en-GB', options).format(date);
 }
 
@@ -107,19 +117,43 @@ function loadData() {
                         ;
                     }
                 }
+
+                $('.last-update').html(formatDatetime((new Date()).getTime() / 1000));
             }
         });
     }
 }
 
-loadData();
+function checkStatus() {
+    let up = $('.status-2').length;
+    let down = $('.status-9').length;
+    let paused = $('.status-1').length;
+    let unknown = $('.status').length - up - down - paused;
+
+    $('.monitors-up').html(up);
+    $('.monitors-down').html(down);
+    $('.monitors-paused').html(paused);
+    $('.monitors-unknown').html(unknown);
+
+    if (down > 0) {
+        $('header').css('background', 'var(--monitor-down)');
+        $('link[rel="icon"]').attr('href', '/favicon-down.ico');
+        $('meta[name="theme-color"]').attr('content', '#D32F2F');
+    } else {
+        $('header').css('background', 'var(--monitor-up)');
+        $('link[rel="icon"]').attr('href', '/favicon.ico');
+        $('meta[name="theme-color"]').attr('content', '#8BC34A');
+    }
+}
+
 let timeout = refresh;
 setInterval(function () {
+    checkStatus();
     if (--timeout === 0) {
         loadData();
+        $('.countdown').html(timeout);
         timeout = refresh;
-        $('#countdown').html(timeout);
     } else {
-        $('#countdown').html(timeout);
+        $('.countdown').html(timeout);
     }
 }, 1000);